Create Layout facade and re-arrenge files so that base layout class called "Layout" instead of "Base"
This fixes Docblocks and class hierarchy
