+++ install-stubs/Orchid/Layouts/Role/RoleEditLayout.php
@@ -12,8 +12,6 @@ class RoleEditLayout extends Rows
     /**
      * Views.
      *
-     * @throws \Throwable|\Orchid\Screen\Exceptions\TypeException
-     *
      * @return array
      */
     public function fields(): array
