+++ install-stubs/Orchid/Screens/Examples/ExampleCardsScreen.php
@@ -7,12 +7,12 @@ use Illuminate\Http\Request;
 use Orchid\Screen\Action;
 use Orchid\Screen\Actions\Button;
 use Orchid\Screen\Contracts\Cardable;
-use Orchid\Screen\Layout;
 use Orchid\Screen\Layouts\Card;
 use Orchid\Screen\Layouts\Compendium;
 use Orchid\Screen\Layouts\Facepile;
 use Orchid\Screen\Screen;
 use Orchid\Support\Color;
+use Orchid\Support\Facades\Layout;
 use Orchid\Support\Facades\Toast;
 
 class ExampleCardsScreen extends Screen
