+++ install-stubs/Orchid/Screens/Role/RoleListScreen.php
@@ -60,7 +60,7 @@ class RoleListScreen extends Screen
     /**
      * Views.
      *
-     * @return Layout[]
+     * @return \Orchid\Screen\Layout[]
      */
     public function layout(): array
     {
