+++ install-stubs/Orchid/Screens/User/UserEditScreen.php
@@ -15,8 +15,8 @@ use Orchid\Screen\Actions\Button;
 use Orchid\Screen\Actions\DropDown;
 use Orchid\Screen\Actions\ModalToggle;
 use Orchid\Screen\Fields\Password;
-use Orchid\Screen\Layout;
 use Orchid\Screen\Screen;
+use Orchid\Support\Facades\Layout;
 use Orchid\Support\Facades\Toast;
 
 class UserEditScreen extends Screen
@@ -93,9 +93,7 @@ class UserEditScreen extends Screen
     }
 
     /**
-     * @throws \Throwable
-     *
-     * @return Layout[]
+     * @return \Orchid\Screen\Layout[]
      */
     public function layout(): array
     {
