+++ src/Screen/Layouts/Metric.php
@@ -5,12 +5,13 @@ declare(strict_types=1);
 namespace Orchid\Screen\Layouts;
 
 use Illuminate\Contracts\View\Factory;
+use Orchid\Screen\Layout;
 use Orchid\Screen\Repository;
 
 /**
  * Class Metric.
  */
-abstract class Metric extends Base
+abstract class Metric extends Layout
 {
     /**
      * @var string
