+++ src/Screen/Layouts/Modal.php
@@ -5,12 +5,13 @@ declare(strict_types=1);
 namespace Orchid\Screen\Layouts;
 
 use Orchid\Screen\Commander;
+use Orchid\Screen\Layout;
 use Orchid\Screen\Repository;
 
 /**
  * Class Modal.
  */
-class Modal extends Base
+class Modal extends Layout
 {
     use Commander;
