+++ src/Screen/Layouts/Table.php
@@ -5,13 +5,14 @@ declare(strict_types=1);
 namespace Orchid\Screen\Layouts;
 
 use Illuminate\Contracts\View\Factory;
+use Orchid\Screen\Layout;
 use Orchid\Screen\Repository;
 use Orchid\Screen\TD;
 
 /**
  * Class Table.
  */
-abstract class Table extends Base
+abstract class Table extends Layout
 {
     /**
      * @var string
