+++ src/Screen/Screen.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use Orchid\Platform\Http\Controllers\Controller;
-use Orchid\Screen\Layouts\Base;
+use Orchid\Screen\Layout;
 use Orchid\Support\Facades\Dashboard;
 use ReflectionClass;
 use ReflectionException;
@@ -69,7 +69,7 @@ abstract class Screen extends Controller
     /**
      * Views.
      *
-     * @return Layout[]
+     * @return \Orchid\Screen\Layout[]
      */
     abstract public function layout(): array;
 
@@ -80,7 +80,7 @@ abstract class Screen extends Controller
      */
     public function build()
     {
-        return Layout::blank([
+        return LayoutFactory::blank([
             $this->layout(),
         ])->build($this->source);
     }
@@ -102,12 +102,12 @@ abstract class Screen extends Controller
         $query = $this->callMethod($method, request()->all());
         $source = new Repository($query);
 
-        /** @var Base $layout */
+        /** @var Layout $layout */
         $layout = collect($this->layout())
             ->map(function ($layout) {
                 return is_object($layout) ? $layout : app()->make($layout);
             })
-            ->map(function (Base $layout) use ($slug) {
+            ->map(function (Layout $layout) use ($slug) {
                 return $layout->findBySlug($slug);
             })
             ->filter()
