+++ tests/Unit/Screen/Layouts/ComponentTest.php
@@ -2,7 +2,7 @@
 
 namespace Orchid\Tests\Unit\Screen\Layouts;
 
-use Orchid\Screen\Layout;
+use Orchid\Screen\LayoutFactory;
 use Orchid\Screen\Repository;
 use Orchid\Tests\App\Components\Hello;
 use Orchid\Tests\TestUnitCase;
@@ -15,7 +15,7 @@ class ComponentTest extends TestUnitCase
             'name' => 'Alexandr Chernyaev',
         ]);
 
-        $layout = Layout::component(Hello::class);
+        $layout = LayoutFactory::component(Hello::class);
 
         $html = $layout->build($repository)->withErrors([])->render();
